requirements.txt
pymorphy2==0.9.1
sentence-transformers
faiss-cpu
networkx
neo4j
umap-learn
hdbscan
numpy
pandas
